==> main_analyses/SCRIPTS/14_interaction_plots_report.Rmd <==
---
title: "Interaction plots: `any_AED` and `PRS`"
author: "Julia Romanowska"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    code_folding: "hide"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(readr)
library(tidyr)
library(dplyr)
library(purrr)
library(here)
library(ggplot2)
```

## Intro

Checking whether there is an interaction between `any_AED` (use of AED) and PRS.

The analyses were done separately for each outcome:

- language impairment,
- autistic trait,

for each PRS type:

- `prs_auto`,
- `prs_manual`, and
- `minor_al_load` (only two strata),

and separately for all the age categories within each outcome.

```{r read_data}
interact_lang_imp_list <- readRDS(here("RESULTS", "interact_lang_imp_list.rds"))
interact_autist_list <- readRDS(here("RESULTS", "interact_autist_list.rds"))
```

## Plots

### Language impairment

```{r}
cur_outcome <- "lang. impair."
cur_list <- interact_lang_imp_list
```

```{r plot_all}
walk(
  names(cur_list), # first level: age categories
  function(cur_age){
    cat("--- CURRENT AGE: ", cur_age, "---\n")
    walk(
      names(cur_list[[cur_age]]), # second level: PRS type
      function(cur_prs_type){
        cat("------ CURRENT PRS TYPE: ", cur_prs_type, "------\n")
        cur_data <- cur_list[[cur_age]][[cur_prs_type]]
        print(
          knitr::kable(
            broom::tidy(cur_data$glm_run)
          )
        )
        print(
          cur_data$plot +
            theme_minimal() +
            labs(
              title = "Interaction plot (any_AED*PRS)",
              subtitle = paste0("outcome: ", cur_outcome, ", age cat.: ", cur_age)
            )
        )
      }
    )
    cat("\n")
  }
)
```

### Autistic traits

```{r}
cur_outcome <- "aut. traits"
cur_list <- interact_autist_list
```

```{r plot_all, fig.path="figure-html/new"}
```
